update to spring boot 2 + rewrite test case + rename

+++ spring-ws-client-timeout/src/main/java/com/codenotfound/ws/endpoint/TicketAgentEndpoint.java
@@ -15,16 +15,19 @@ import org.springframework.ws.server.endpoint.annotation.ResponsePayload;
 @Endpoint
 public class TicketAgentEndpoint {
 
-  @PayloadRoot(namespace = "http://example.org/TicketAgent.xsd", localPart = "listFlightsRequest")
+  @PayloadRoot(namespace = "http://example.org/TicketAgent.xsd",
+      localPart = "listFlightsRequest")
   @ResponsePayload
   public JAXBElement<TFlightsResponse> listFlights(
-      @RequestPayload JAXBElement<TListFlights> request) throws InterruptedException {
+      @RequestPayload JAXBElement<TListFlights> request)
+      throws InterruptedException {
 
     // sleep for 10 seconds so a timeout occurs
-    Thread.sleep(10 * 1000);
+    Thread.sleep(10 * (long) 1000);
 
     ObjectFactory factory = new ObjectFactory();
-    TFlightsResponse tFlightsResponse = factory.createTFlightsResponse();
+    TFlightsResponse tFlightsResponse =
+        factory.createTFlightsResponse();
     tFlightsResponse.getFlightNumber().add(BigInteger.valueOf(101));
 
     return factory.createListFlightsResponse(tFlightsResponse);
